image_classification/requirements-gpu.txt
tensorflow-gpu
tqdm
colorama
